requirements.txt
asgiref==3.7.2
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
